Updated files to compile under Ubuntu Natty.
git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp-misc/trunk/rtpaudio2@1020 acdd8558-d4fd-0310-b0a5-8cf3c1ec824a

+++ rtpaudio/client.cc
@@ -376,9 +376,9 @@ int main(int argc, char* argv[])
       const cardinal currentLayers = client->getLayers();
       const card64 now = getMicroTime();
       if(!optAudioDebug) {
-         printf("\x0d%2Ld:%02Ld.%02Ld   [Quality: %d Hz / %d Bit / %s]  [%s]      ",
-                (seconds / 60),(seconds % 60),
-                (position % PositionStepsPerSecond) / (PositionStepsPerSecond / 100),
+         printf("\x0d%2u:%02u.%02u   [Quality: %d Hz / %d Bit / %s]  [%s]      ",
+                (unsigned int)(seconds / 60), (unsigned int)(seconds % 60),
+                (unsigned int)((position % PositionStepsPerSecond) / (PositionStepsPerSecond / 100)),
                 client->getSamplingRate(),
                 client->getBits(),
                 ((client->getChannels() == 2) ? "Stereo" : "Mono"),
